Rewrite parallel DD based on futures
Use thread-based futures for parallelism. Theoretically, we could
suffer a performance hit caused by the GIL, if tests were performed
in-process, but in practice, tests are usually executed in/by
subprocesses anyway. Thus, it is superfluous to start a subprocess
only to start up another subprocess.
