+++ picire/__init__.py
@@ -17,6 +17,5 @@ from .iterator import CombinedIterator, IteratorRegistry
 from .limit_reduction import LimitReduction
 from .outcome import Outcome
 from .parallel_dd import ParallelDD
-from .shared_cache import shared_cache_decorator
 from .splitter import SplitterRegistry
 from .subprocess_test import ConcatTestBuilder, SubprocessTest
