+++ setup.cfg
@@ -30,7 +30,6 @@ install_requires =
     chardet
     importlib-metadata; python_version < "3.8"
     inators
-    psutil < 5.9.6  # FIXME: PyPy does not expose PyErr_SetExcFromWindowsErrWithFilenameObject on Windows, waiting for psutil workaround
 
 [options.entry_points]
 console_scripts =
